+++ src/main/java/com/wafersystems/virsical/map/controller/ParkController.java
@@ -6,9 +6,10 @@ import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.wafersystems.virsical.common.core.util.R;
 import com.wafersystems.virsical.map.common.BaseController;
-import com.wafersystems.virsical.map.common.MsgConstants;
+import com.wafersystems.virsical.map.common.MapMsgConstants;
 import com.wafersystems.virsical.map.entity.Building;
 import com.wafersystems.virsical.map.entity.Park;
+import com.wafersystems.virsical.map.service.IBuildingService;
 import com.wafersystems.virsical.map.service.IFloorService;
 import com.wafersystems.virsical.map.service.IParkService;
 import io.swagger.annotations.Api;
@@ -40,13 +41,15 @@ public class ParkController extends BaseController {
 
   private final IFloorService floorService;
 
+  private final IBuildingService buildingService;
+
   @ApiOperation(value = "添加园区", notes = "添加园区")
   @ApiImplicitParam(name = "park", value = "园区对象", required = true, dataType = "Park")
   @PostMapping("/add")
   public R add(@RequestBody Park park) {
     Park one = parkService.getOne(Wrappers.<Park>query().lambda().eq(Park::getParkName, park.getParkName()));
     if (one != null) {
-      return R.fail(MsgConstants.PARK_NAME_NO_REPEAT);
+      return R.fail(MapMsgConstants.PARK_NAME_NO_REPEAT);
     }
     return parkService.save(park) ? R.ok() : R.fail();
   }
@@ -58,7 +61,7 @@ public class ParkController extends BaseController {
     Park one = parkService.getOne(Wrappers.<Park>query().lambda().eq(Park::getParkName, park.getParkName())
         .ne(Park::getParkId, park.getParkId()));
     if (one != null) {
-      return R.fail(MsgConstants.PARK_NAME_NO_REPEAT);
+      return R.fail(MapMsgConstants.PARK_NAME_NO_REPEAT);
     }
     return parkService.updateById(park) ? R.ok() : R.fail();
   }
@@ -67,6 +70,10 @@ public class ParkController extends BaseController {
   @ApiImplicitParam(name = "id", value = "园区id", required = true, dataType = "Integer")
   @PostMapping("/delete/{id}")
   public R delete(@PathVariable Integer id) {
+    int count = buildingService.count(Wrappers.<Building>query().lambda().eq(Building::getParkId, id));
+    if (count > 0) {
+      return R.fail(MapMsgConstants.THIS_PARK_HAS_BUILDING);
+    }
     return parkService.removeById(id) ? R.ok() : R.fail();
   }
